+++ CHANGES.md
@@ -5,6 +5,7 @@
 * Fix potential race condition when reopening the file (issue 5)
 * Fix potential blocking of `tail.Stop` (issue 4)
 * Fix uncleaned up ChangeEvents goroutines after calling tail.Stop
+* Support Follow=false
 
 # Feb, 2013